Truncate File basic requirements are done
I still need to work on stretch goals. My HTML now says my name and
Truncate file has my name in it.

I added an experiment file to see if I can get the same result of
truncating 8 words, but with a more optimal approach. If I can get it to
work then I will return it to the html as well.